code improvements

+++ src/main/java/com/xresch/cfw/validation/FileCanWriteValidator.java
@@ -7,8 +7,9 @@ import com.xresch.cfw.utils.Ternary;
 
 /**************************************************************************************************************
  * 
- * @author Reto Scheiwiller, (c) Copyright 2019 
- * @license Creative Commons: Attribution-NonCommercial-NoDerivatives 4.0 International
+ * @author Reto Scheiwiller, (c) Copyright 2019
+ * @license Creative Commons: Attribution-NonCommercial-NoDerivatives 4.0
+ *          International
  **************************************************************************************************************/
 public class FileCanWriteValidator extends AbstractValidator {
 
@@ -19,39 +20,40 @@ public class FileCanWriteValidator extends AbstractValidator {
 
 	@Override
 	public boolean validate(Object value) {
-		
+
 		Ternary result = validateNullEmptyAllowed(value);
-		if(result != Ternary.DONTCARE ) return result.toBoolean();
+		if (result != Ternary.DONTCARE)
+			return result.toBoolean();
 
 		File file = null;
-		
-		if(value instanceof String) {
-			file = new File((String)value);
-		}else if(value instanceof File) {
-			file = (File)value;
-		}else{
-			this.setInvalidMessage("Unsupported type for FileCanWriteValidator: '"+value.getClass().getName()+"'");
-		}
-		
 
-		if(file.exists()){
-			if(!file.canWrite()){
-				this.setInvalidMessage("File cannot be written: '"+file.getAbsolutePath()+"'");
+		if (value instanceof String) {
+			file = new File((String) value);
+		} else if (value instanceof File) {
+			file = (File) value;
+		} else {
+			this.setInvalidMessage("Unsupported type for FileCanWriteValidator: '" + value.getClass().getName() + "'");
+			return false;
+		}
+
+		if (file != null && file.exists()) {
+			if (!file.canWrite()) {
+				this.setInvalidMessage("File cannot be written: '" + file.getAbsolutePath() + "'");
 				return false;
 			}
-		}else{
+		} else {
 			boolean success = true;
 			try {
 				success &= file.mkdirs();
 				success &= file.createNewFile();
 				success &= file.delete();
 			} catch (IOException e) {
-				this.setInvalidMessage("File cannot be written: '"+file.getAbsolutePath()+"'");
+				this.setInvalidMessage("File cannot be written: '" + file.getAbsolutePath() + "'");
 				return false;
 			}
 			return success;
 		}
-		
+
 		return false;
 	}
 }
